--- Util.h
@@ -1,16 +1,0 @@
-//
-// Created by ASUS on 18/5/2024.
-//
-
-#ifndef B8_SANDIMORSE_UTIL_H
-#define B8_SANDIMORSE_UTIL_H
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <ctype.h>
-
-//make void getline
-
-
-#endif //B8_SANDIMORSE_UTIL_H
